Divest from mage
I don't think we're doing enough with mage to warrant its existence. Our
CI is already split between mage tasks and vanilla go commands. As are
our local dev instructions (or at least the ones provided to
CLAUDE.md).

I think the biggest benefit it is providing is applying the -ldsflags for
us. But I think we should be equally fine hardcoding a version and
overwriting it at build/release time?