Add a git client that can get the current branch
We'll start including this as part of the event payload when triggering
runs from the CLI via `rwx run`.

I've included a few test fixtures that are just shell scripts that set
up repos the way I want them to look. This was after a bit of
experimenting with doing it all in go, adding git submodules, and using
git bundles. But in the end, I figured this was the easiest to
understand and extend for more complicated scenarios.

Nothing uses this client yet.