Find the common ancestor git commit
We want to begin supporting the ability to patch a remote run with a git
patch. To do so, we need to identify a point of common ancestry - the
latest commit that is likely to exist both remotely and locally -
because that's the commit we'll apply the patch against.

The algorithm I've landed on for now is:

```
git for-each-ref --format "%(objectname) %(refname)" refs/remotes
```

Which has output like:

```
18d5404642086cff539c1e3b639531aca2e03766 refs/remotes/origin/HEAD
241a6872fbb1294965abce116e2d133aeb76bf4d refs/remotes/origin/dan/notary-list
...
```

In short, it gives us the HEAD commit for every known remote ref. This
is only as recent as the last `git fetch` which we're intentionally not
running so as not to have our lookup mutate local state.

And then we walk backward on our current ref to find the first commit
that matches.

The alternative is to use `git merge-base` but that requires us to best-effort
identify the remote "default" ref (what if they don't use the `origin`
convention?)