Introduce GeneratePatchFile
This adds an opinionated GeneratePatchFile method to our git client. It
considers the following to be relevant parts of a patch:

- committed changes (the behavior you'd usually get if you did
  `git format-patch`)
- uncommitted changes to tracked files (the behavior you get if you run
  `git diff`)
- changes to binary files (the `--binary` flag)

It does not consider untracked files to be part of the patch.

We'll eventually use this method to support patching CLI runs. Like the
other git methods I've added, it fails open - if it can't do something,
it just returns empty. This is because we don't want this behavior to
get in the way of launching runs if there's a problem.

The tests follow the same shell script fixture pattern I introduced in:
https://github.com/rwx-cloud/cli/pull/196

This means the commits are non-deterministic (I briefly looked into
making the fixtures produce deterministic commits but it wasn't worth
the complexity) so the patches are non-deterministic. Hence the pretty
loose string contains assertions I'm making.