Introduce mock git client
While starting to wire everything together, I decided it'd be nicer if
our git client followed the same mocking strategy as our API and SSH
clients.

So this sets all that up.

Now, the only tests that actual rely on real git interactions are the
tests of the git client itself. Everything else can just be mocked. Yay.