Remove the CI flag (at least for now)
Let's keep this simple for now so we can get to passing tests. There
might still be a valid reason to add this, but the other opt-out flag
might be enough.

And there's a bit of a challenge of making sure our own CI can run the
tests which requires a bit more care around what env gets passed into
these tests.