Only find common ancestors from origin
The problem remains the same: we want to identify a common ancestor on a
likely remote. Previously, this was checking _any_ remote so if you
follow a different naming convention or have multiple remotes
configured, it might pick a commit sha that exists on a remote that's
different than the one you're cloning in rwx.

This plays the odds and just assumes the conventional origin. Otherwise,
it'll fail open. We can figure out the best way to handle a different
remote name when we first encounter that situation - presumably by
adding some additional configuration?

The test was actually originally named to do this - but I goofed and was
returning stdout from my test fixture so the assertion was wrong. I've
made all the empty tests more explicit to avoid this class of oopsie in
the future.