Hide git patching behind a preview feature flag
There's still some UX polish that needs to happen around this feature
but I want to make sure releasing the CLI remains unblocked if we need
to do that before this work wraps up.

So this introduces a temporary feature flag that makes this opt-in.
Which should return main to a releasable state.