Support dispatch triggers
When I went to generate the CLI trigger for .rwx/release.yml, it was a
noop.

This is because the original implementation narrowly fit for the shape
of github triggers:

```
on:
  github:
    push:
      init:
        ref: ${{ event.git.ref }}
```

Which is of type `MappingNode`.

But dispatch triggers are structured:

```
on:
  dispatch:
    - key: my-dispatch
      init:
        ref: ${{ event.git.ref }}
```

Which is of type `SequenceNode` (an array of other nodes).

Also I noticed a bug where I was only allowing params that matched the
name of the valid git event keys.