Accept file as positional argument
This suggests a change to `rwx run` to accept `--file/-f` as a required
first positional argument. This is because every invocation requires a
file. `--file/-f` are still accepted to maintain backward compabitility.

Previously, positional arguments were interpreted as targeted tasks.
This also suggests a breaking change that forces them to be supplied
with `--task` flags. This is because task targeting is infrequently
included in `rwx run` invocations so there's no reason they should be
special cased as positional arguments.

This means commands like:

```
rwx run --file ci.yml
```

will continue to work. But can be rewritten as:

```
rwx run ci.yml
```

Commands like:

```
rwx run --file ci.yml task1 task2
```

will now fail and should be rewritten as:

```
rwx run ci.yml --task task1 --task task2
```

An alternate approach I considered but did not choose is to still allow
the tasks as positional arguments if `--file/-f` is supplied. It would
allow it to be a fully backward compatible change in exchange for
additional CLI parsing complexity.