Add `rwx image pull`
This is just the back half of `rwx image build` - pull and tag without
launching the run. Now that any task on RWX can be represented as a
container image, this makes it easier to pull it.

And make some copy a bit more consistent across the various `rwx image`
help text.