Fixup merge conflict on #236 and #237
I was too trigger happy in merging and my test that I introduced in #237
used the old API from before #236.

I also lost a commit that added CLI validation against passing
`--no-pull` and `--tag` together so I've added that back. This was just
user error on my part (forgetting to push) and not related to the merge
conflict.