Use git merge-base to find common ancestor
Originally, https://github.com/rwx-cloud/cli/pull/196 went to lengths to
work even in cases where the convention of naming the remote `origin`
was not followed. So we used `git for-each-ref` to get the `HEAD` of
every known remote ref and then we walk our local history to find a
match.

But https://github.com/rwx-cloud/cli/pull/202 walked this back - it
assumes `origin` is used. So we can move to the simpler `git
merge-base`.

This should help in the case where you're in detached `HEAD` locally and
the commit is no longer the `HEAD` of any remote ref.