When in detached HEAD, send current commit
An alternative to https://github.com/rwx-cloud/cli/pull/252 that aims to
solve the underlying edge case (finding a commit to send when a specific
commit is checked out locally) without changing how we find the commit
to send generally.

Basically, we want to try to avoid anything that will mutate just to do
this lookup - issuing pulls or fetches or what not on the user's behalf
- and we want to avoid anything that's provider specific - using the
github api or similar instead of just git. On the happy path, we work
within these constraints as described in
https://github.com/rwx-cloud/cli/pull/196 to find the remote refs we
already know about and see where we diverge from them in our history.

But when checking out a specific commit and ending up in detached HEAD
state, we can end up in a pretty distant point of divergence. For
example, a really long lived feature branch that eventually merges via a
squash merge. This instead proposes that when a specific commit is
checked out, it declares intent that this is the commit we're
interacting with. And if we're going to git patch, we should just try to
git patch against it rather than try to look up a potentially distant
merge base. Generally, one of two things will be true:

- the commit will exist on a remote in this state: "This commit does not
  belong to any branch on this repository, and may belong to a fork
  outside of the repository."
- the commit has only been created locally and then has been checked out
  directly - in which case the user can push it.