Exclude .rwx directory from git patch
The .rwx directory is already made available to the task separately.

When it's included in the patch, it can cause unfiltered tasks to cache
bust more than they should while iterating on workflow changes.

Git commands allow a pathspec to include/exclude certain poaths. This
commit modifies our git client to allow the caller to pass a pathspec
when generating a patch file.

And then our calling service passes `. :!.rwx` (or more accurately,
whatever the .rwx directory happens to be set to) to exclude the
directory with the workflow files.

Before (preview the patch):

https://cloud.rwx.com/mint/rwx/runs/2cb2ddcc1d64495d8fe82bd49e69187f

After (preview the patch):

https://cloud.rwx.com/mint/rwx/runs/f3fdb06251bd495ea4d372374a8911d7