Add go fmt to CI
This also unearthed a few places where cache busting (because
.rwx/continous_integration.yml` changed in a git patch) was causing
issues in `go mod tidy` and in a unit test that both rely on `git diff`.

The underlying cause of the cache bust is addressed in:

https://github.com/rwx-cloud/cli/pull/259

And this contains a few fixes to make sure even if `git diff` is called
after a git patch is applied, we'll be fine. In the go tooling, by
capturing the diff beforehand and then comparing. In the test, by
running it in a different directory.